Fix passing multiple `lint_warnings` to `buildifier`
Buildifier accepts a single `--warning` flag with a comma-separated
list of warnings to enable/disable instead of a repeated flag usage.

Also removes mentions of `all - some` warnings in examples as Buildifier
doesn't support this. Due to the bug fixed by this PR, these settings
previously resulted in `default - some`.