bash: use -prune in find (#127)
Use -prune to handle excluded paths, rather than simply ! -path, which
will still visit all subdirectories and files in the excluded paths.

This is more efficient, but more importantly avoids visiting files which
may not be readable by the user, which causes `find` fail, and causes
buildifier to fail as we are using `-o pipefail`.